+++ CHANGELOG.rst
@@ -4,6 +4,21 @@ CHANGELOG
 Development
 -----------
 
+- [Changed] Local sky model endpoint updated.
+  (`MR118 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/118>`__)
+
+  - CSV downloads by addding format=csv as query parameter.
+  - LSM table has pagination.
+- [Update] Catch issue when 2 versions are created for the same catalogue at the same time
+  (`MR117 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/117>`__)
+- [Change] Update LocalSkyModelQuery to do:
+  (`MR116 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/116>`__)
+
+  - Allow HTTP endpoint to return multiple catalogues
+  - Do not allow Flow based search to use multiple catalogues
+  - Make Flow based LSM be forced to query by ``catalogue_name``
+  - Update HTML page to be clear when no components match the query
+  - Upgrade fastapi+uvicorn
 - [New] new commands to cleanup catalogues
   (`MR106 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/106>`__)
 - [Update] Redirect from '/' to '/docs' Improve swagger example code.
@@ -12,10 +27,8 @@ Development
   (`MR105 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/105>`__)
 - [Remove] Unused dependencies from pyproject.toml
   (`MR105 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/105>`__)
-
 - [Deprecate] HTML upload page is deprecated
   (`MR105 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/105>`__)
-
 - [Update] Improve error reporting during CSV file validation
   (`MR108 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/108>`__)
 - [Fix] Allow only a single metadata file to be uploaded